+++ weyl-frontend/app/components/SelectionPane/index.js
@@ -1,18 +1,15 @@
 import React from "react";
 import {
-  Button,
   Classes,
-  Collapse,
-  Overlay,
   Spinner,
 } from "@blueprintjs/core";
-import classNames from "classnames";
-import naturalsort from "javascript-natural-sort";
 import * as $ from "jquery";
 import * as _ from "underscore";
+import naturalsort from "javascript-natural-sort";
 
 import Pane from "../Pane";
 import NonHistograms from "./NonHistograms";
+import Histograms from "./Histograms";
 import { defaultTitle, curatedTitles } from "./behaviors";
 
 class SelectionPane extends React.Component { // eslint-disable-line react/prefer-stateless-function
@@ -41,19 +38,32 @@ class SelectionPane extends React.Component { // eslint-disable-line react/prefe
         visible={visible}
         onClose={this.props.onClose}
       >
-        {
-          (histogramEnabled(entityIds))
-            ? <Histograms histograms={histograms} />
-            : (
-              <NonHistograms
-                featureAttributes={attributes}
-                behavior={getBehavior(singleLayer(entityIds, layers))}
-              />
-            )
+        { (loaded) ?
+          <SelectionView
+            entityIds={entityIds}
+            histograms={histograms}
+            attributes={attributes}
+            layers={layers}
+          />
+        : <div style={{ width: "100%", textAlign: "center"}}>
+            <Spinner className={Classes.LARGE} />
+          </div>
         }
       </Pane>
     );
   }
+}
+
+const SelectionView = ({entityIds, histograms, attributes, layers}) => {
+    if (histogramEnabled(entityIds)) {
+      return <Histograms histograms={histograms} />
+    }
+    return (
+      <NonHistograms
+        featureAttributes={attributes}
+        behavior={getBehavior(singleLayer(entityIds, layers))}
+      />
+  );
 }
 
 // entityIds is an object { layerId -> [entityIds] }
@@ -64,56 +74,7 @@ const numEntities = (entityIds) => _.size(_.flatten(_.values(entityIds)));
 
 const singleLayer = (entityIds, layers) => layers[_.keys(entityIds)[0]];
 
-// TODO: blessed vs. non-blessed
-const Histograms = ({ histograms }) => (
-  <div style={{ maxHeight: "30em", overflow: "auto" }}>
-    <table className="ui celled very compact small fixed selectable table">
-      {
-        _.chain(histograms)
-          .sort((a, b) => naturalsort(a.attribute, b.attribute))
-          .map(histogram =>
-            <AttributeHistogram
-              key={histogram.attribute}
-              histogram={histogram}
-            />
-          )
-          .value()
-      }
-    </table>
-  </div>
-);
 
-const AttributeHistogram = ({ histogram }) => (
-  <tbody className="ui accordion" ref={x => $(x).accordion()}>
-    <tr className="title">
-      <td style={{ fontWeight: "bold" }}>
-        <i className="dropdown icon"></i>
-        {histogram.attribute}
-      </td>
-    </tr>
-
-    <tr className="content">
-      <td>
-        <table className="ui celled very compact small fixed selectable definition table">
-          <tbody>
-            {
-              _.chain(histogram.buckets)
-                .sort((a, b) => naturalsort(a.value, b.value))  // Fall back to alphabetical
-                .sort((a, b) => b.count - a.count)              // Highest count first
-                .map(bucket =>
-                  <tr key={bucket.value}>
-                    <td className="right aligned">{bucket.value}</td>
-                    <td>{bucket.count}</td>
-                  </tr>
-                )
-                .value()
-            }
-          </tbody>
-        </table>
-      </td>
-    </tr>
-  </tbody>
-);
 
 const getBehavior = (layer) => {
   const layerName = layer.metadata.name;
